ks - added condition to Score DAO methods to set scores to -1 when null.  passing -1 to backing beans for "N/A".  changed DriverReportItem and VehicleReportItem to use Integers for scores instead of floats.  updated JSF to display scores with ScoreConverter.

+++ dao/src/main/java/com/inthinc/pro/model/DriverReportItem.java
@@ -7,10 +7,10 @@ public class DriverReportItem extends BaseEntity
     private String employee;
     private Vehicle vehicle;
     private Integer milesDriven;
-    private Float overallScore;
-    private Float speedScore;
-    private Float styleScore;
-    private Float seatBeltScore;
+    private Integer overallScore;
+    private Integer speedScore;
+    private Integer styleScore;
+    private Integer seatBeltScore;
     private String styleOverall;
     private String styleSpeed;
     private String styleStyle;
@@ -59,35 +59,35 @@ public class DriverReportItem extends BaseEntity
     {
         this.milesDriven = milesDriven;
     }
-    public Float getOverallScore()
+    public Integer getOverallScore()
     {
         return overallScore;
     }
-    public void setOverallScore(Float overallScore)
+    public void setOverallScore(Integer overallScore)
     {
         this.overallScore = overallScore;
     }
-    public Float getSpeedScore()
+    public Integer getSpeedScore()
     {
         return speedScore;
     }
-    public void setSpeedScore(Float speedScore)
+    public void setSpeedScore(Integer speedScore)
     {
         this.speedScore = speedScore;
     }
-    public Float getStyleScore()
+    public Integer getStyleScore()
     {
         return styleScore;
     }
-    public void setStyleScore(Float styleScore)
+    public void setStyleScore(Integer styleScore)
     {
         this.styleScore = styleScore;
     }
-    public Float getSeatBeltScore()
+    public Integer getSeatBeltScore()
     {
         return seatBeltScore;
     }
-    public void setSeatBeltScore(Float seatBeltScore)
+    public void setSeatBeltScore(Integer seatBeltScore)
     {
         this.seatBeltScore = seatBeltScore;
     }
